python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tensorflow>=2.15.0",
    "tensorflow-metal>=1.1.0",
    "tensorflowjs>=4.17.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.4.0",
    "Pillow>=10.2.0",
    "ipykernel==6.30.1",
    "kagglehub>=0.2.0",
    "datasets>=4.3.0",
    "marimo>=0.17.2",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.2.0",
]
